+++ tests/helpers/testHelper.h
@@ -2,15 +2,8 @@
 // Copyright Matthias Akstaller 2019 - 2024
 // MIT License
 
-#ifndef MO_CPP_CONSOLE_OUT
-#define MO_CPP_CONSOLE_OUT
-
-/**
-* Prints a string to the c standart console
-*
-* @param msg pointer to the string
-*/
-void cpp_console_out(const char *msg);   
+#ifndef MO_TESTHELPER_H
+#define MO_TESTHELPER_H
 
 extern unsigned long mtime;
 unsigned long custom_timer_cb();