+++ tests/ocppEngineLifecycle.cpp
@@ -10,9 +10,6 @@
 TEST_CASE( "Context lifecycle" ) {
     printf("\nRun %s\n",  "Context lifecycle");
 
-    //set console output to the cpp console to display outputs
-    //mocpp_set_console_out(cpp_console_out);
-
     //initialize Context with dummy socket
     MicroOcpp::LoopbackConnection loopback;
     mocpp_initialize(loopback);